use jwt in cookies changes

+++ server/src/main/java/com/madebywael/zhonya/config/JwtService.java
@@ -23,7 +23,7 @@ public class JwtService {
     @Value("${application.security.jwt.secret-key}")
     private String SECRET_KEY;
     @Value("${application.security.jwt.expiration}")
-    private long jwtExpiration;
+    private int jwtExpiration;
 
     public String extractUserEmail(String token) {
         return extractClaim(token, Claims::getSubject);
